+++ src/main/java/si/matjazcerkvenik/alertmonitor/providers/EventloggerDataProvider.java
@@ -17,6 +17,7 @@ package si.matjazcerkvenik.alertmonitor.providers;
 
 import si.matjazcerkvenik.alertmonitor.model.DEvent;
 import si.matjazcerkvenik.alertmonitor.model.DTarget;
+import si.matjazcerkvenik.alertmonitor.model.DWarning;
 import si.matjazcerkvenik.alertmonitor.util.LogFactory;
 import si.matjazcerkvenik.alertmonitor.web.WebhookMessage;
 
@@ -27,7 +28,7 @@ public class EventloggerDataProvider extends AbstractDataProvider {
     @Override
     public void init() {
         logger.info(providerConfig.toString());
-        addWarning("notSupportedProvider", "Provider not supported");
+        addWarning("notSupportedProvider", "Provider not supported", DWarning.DWARNING_SEVERITY_WARNING);
     }
 
     @Override